replace LegacyRef, move types, clean imports, onLayoutChange > inner function
